Bug-fix in start_ts() function of UPID
Summary: There was an incorrect static_cast in the UPID class that used uint32_t instead of int64_t.

Test Plan: Updated the test to use larger timestamps. That catches the issue.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1801

GitOrigin-RevId: 05ec88524a5cd7d9278d9ebab4ab8e83c9883ed8

+++ src/shared/metadata/base_types.h
@@ -52,7 +52,7 @@ class UPID {
 
   uint32_t asid() const { return static_cast<uint32_t>(absl::Uint128High64(value_) >> 32U); }
 
-  int64_t start_ts() const { return static_cast<uint32_t>(Uint128Low64(value_)); }
+  int64_t start_ts() const { return static_cast<int64_t>(Uint128Low64(value_)); }
 
   absl::uint128 value() const { return value_; }
